+++ lib/src/arb_to_class_generator.dart
@@ -11,9 +11,8 @@ import 'package:dart_style/dart_style.dart';
 import 'package:gsheet_to_arb/src/arb/arb_generator.dart';
 
 class TranslationsGenerator {
-  void buildTranslations(ArbDocument document, String directory) {
-    var className = "L10n";
-
+  void buildTranslations(ArbDocument document, String directory,
+      String className) {
     var translationClass = Class((ClassBuilder builder) {
       builder.name = className;
       builder.docs.add(
@@ -40,11 +39,28 @@ class TranslationsGenerator {
 
   Method _getFieldGetter(ArbEntry entry) {
     Method method = Method((MethodBuilder builder) {
-      builder.name = "${entry.key}";
+      var context = entry.attributes['context'];
+      var description = entry.attributes['description'];
+      if (description == null) {
+        description = entry.key;
+      }
+
+      var key;
+
+      if (context != null) {
+        key = "${context}_${entry.key}";
+      } else {
+        key = "${entry.key}";
+      }
+
+      builder.name = key;
       builder.type = MethodType.getter;
       builder.lambda = true;
       builder.body =
-          Code("""Intl.message("${entry.value}", name: "${entry.key}")""");
+          Code("""Intl.message("${entry
+              .value}", name: "${key}", desc: "${description}")""");
+      builder.docs.add("\t/// ${description}");
+
     });
     return method;
   }