+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 - Import ARB from Google Sheets
 
 ## 0.0.3
+
+- Generate dart class from arb file